Add serial codes
So part of the way the Atlas is supposed to work is that every person
aboard the Atlas is assigned a serial code, for data registration,
record keeping, and proper fitting as it turns out. This is formatted as
XXXX-YYY, where YYY is the sector the person lives in and XXXX is a
distinct identifier hopefully unique to each person in a sector.
Identifiers are only really retired once its current user has died, and
it's considered traditional, especially amongst nobles, to reuse a
serial code of a deceased ancestor for your firstborn.

I should probably write this in a place that is NOT a commit message,
but that can be done LATER.