Add serial code info
If you saw a text file get committed to this repository and wondered why
the heck it's just a list of numbers, there's a reason for that. And now
I'm trying to add that information into a space that is not a commit
message. Because, ah ha, commit messages are DEFINITELY going to get
read.