Add etcetera file
I'm creating a space to add the various minor ideas I have for the
setting that I'm not ready to add to an existing plot, as well as plot
bunnies and gaps that I would like to see filled. This file may very
well turn into a monster as this project continues.