Add "Angel" story
This came to me last night. I'm not sure exactly WHAT I want to do with
this, but it seemed like something that would be interesting to work
with.

I'll probably work on this with somebody who's more closely in tune with
Sentai tropes and whatnot to make sure I get the proper feel down.