Rework story to address MegaMan
I'm having one of those moments where the characters are making it
difficult to write the story I had in mind. I'm turning into a REAL
writer! Unfortunately it also means I've written myself into a hole.
Oops.

Because I have a few not really good options:
1. Break character to skip over the issue. The easiest, but also
   the least satisfying answer.
2. Have the author roleplay as Lan to address MegaMan's fears, which
   would work, but not something I'm really interested in doing
   narratively, because it undercuts the narrative I'm trying to write.
3. Convince MegaMan that Lan's not coming because he's not real. This
   was the suggestion I got for getting out of this hole, but it
   feels... excessively CRUEL to me.

I'm not sure which option I want to go with, because not of them feel
right to me. They all kinda SUCK.