Add Force Gems story
So it turns out that there are two mainline Zelda games that do not use
the standard currency, the Rupee. One of these is Zelda II which
straight up does not use currency. The other is Four Swords Adventures
which uses a different gem-based currency called Force Gems. I thought
it would be fun to have this end awkwardly.