Add description to forest in Open Hand
I shared this story with somebody new and making the woods seem more of
a marvel was their first suggestion. I thought it was a good decision to
make!

This also includes a bit of reworking of lines since I keep a strict
line limit here. Makes using Vim nicer.