Edit word choice in Vai'ren worldbuilding doc
So, the site I had in mind for the Vai'ren look called their pants
"leggings" as they were primarily designed with fitness in mind.
Unfortunately, the look they were going for was not the look the usual
definition of that word implies. I had that pointed out to me and I
decided it was probably worth the change!