Continue Navi story
I think I'm going to put this story to rest. It's a neat concept, but I
don't think I'm the writer for it. I don't think I have enough
interesting things to do with the concept to warrant going much further
on it. Or maybe it's just that I don't have it in me to write about a
character screaming at reality.